Extract providers into client component
