+++ src/app/page.tsx
@@ -1,0 +1,5 @@
+const HomePage = () => {
+	return <div>Hi there!</div>
+}
+
+export default HomePage
